rename commitrmnocb to commit
