+++ commit/chain_support.go
@@ -1,4 +1,4 @@
-package commitrmnocb
+package commit
 
 import (
 	"fmt"
