+++ commit/factory.go
@@ -1,4 +1,4 @@
-package commitrmnocb
+package commit
 
 import (
 	"context"
