+++ commit/validate_observation.go
@@ -1,4 +1,4 @@
-package commitrmnocb
+package commit
 
 import (
 	"context"
